eventaggregator problem

+++ Interfaces/Database/Datatables/IItemPositionDbt.cs
@@ -9,7 +9,7 @@ namespace Billmanager.Interfaces.Database.Datatables
     public interface IItemPositionDbt : IDatabaseTable
     {
         int BillId { get; set; }
-        int Amount { get; set; }
+        decimal Amount { get; set; }
         decimal Price { get; set; }
         string Description { get; set; }
         decimal Total { get; }
